Update API routes to include versioning and refactor code
This commit updates the API routes across multiple controllers to include a versioning prefix (`v1`), enhancing the API's structure for future updates. Key changes include:

- Updated routes in `CategoriesController`, `QuestionsCategoriesController`, `CollectionsController`, `CollectionsLikesController`, `CollectionsReviewsController`, `CollectionsSubmitionsController`, `HomeScreenController`, `UsersController`, `ChoicesController`, `ChosenChoicesController`, `ExplanationsController`, and `QuestionsController`.
- Added and modified using directives for necessary namespaces.
- Refactored methods for consistency and added authorization checks where needed.
- Improved comments for clarity regarding authorization and user ID handling.
- Introduced new fields in `CollectionsAuthService` and `CollectionsSubmitionsService` to enhance service management.

These changes improve the organization, readability, and maintainability of the codebase while ensuring proper API versioning.

+++ API_Layer/Controllers/Collections/CollectionsSubmitionsController.cs
@@ -1,4 +1,5 @@
 ﻿using API_Layer.Extensions;
+using Core.Authorization_Services.Interfaces;
 using Core.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -6,23 +7,29 @@ using System.Security.Claims;
 
 namespace API_Layer.Controllers.Collections
 {
-    [Route("api/collectionsSubmitions")]
+
+
+    [Route("api/v1/submitions")]
     [ApiController]
     public class CollectionsSubmitionsController : ControllerBase
     {
         ICollectionsSubmitionsService _collectionService;
+        private readonly ICollectionsAuthService _collectionsAuthService;
 
-        public CollectionsSubmitionsController(ICollectionsSubmitionsService collectionService)
+        public CollectionsSubmitionsController(ICollectionsSubmitionsService collectionService, ICollectionsAuthService CollectionsAuthService)
         {
             _collectionService = collectionService;
+            _collectionsAuthService = CollectionsAuthService;
         }
 
         [HttpPost]
         [Authorize]
         public async Task<ActionResult<int>> AddSubmission(int CollectionID)
         {
-
             int UserId = User.GetUserId();
+            if (! await _collectionsAuthService.IsUserCollecOwnerAsync(CollectionID, UserId)) return Unauthorized();
+            
+
             int submissionId = await _collectionService.AddSubmission(CollectionID, UserId);
             return Ok(submissionId);
         }
@@ -33,6 +40,7 @@ namespace API_Layer.Controllers.Collections
         {
             int? userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
             if (userId == null) return Unauthorized();
+            if (!await _collectionsAuthService.IsUserSubmitionOwnerAsync(SubmitionID, (int)userId)) return Unauthorized();
 
             bool isDeleted = await _collectionService.DeleteSubmition(SubmitionID, (int)userId);
 
@@ -47,6 +55,8 @@ namespace API_Layer.Controllers.Collections
         {
             int? userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
             if (userId == null) return Unauthorized();
+
+            //authorization (checking if the sub is the user's) check will make the api slower and it is not really needed here at the moment
 
             var submission = await _collectionService.GetBySubmissionID(SubmissionID, (int)userId);
             if (submission == null) return NotFound();
